+++ src/mos6502/mos6502_inc.c
@@ -16,24 +16,24 @@
 size_t
 mos6502_inc_abs8(int8_t value)
 {
-	return mos6502_write(0xe6) + mos6502_write(value);
+	return mos6502_write((int8_t) 0xe6) + mos6502_write(value);
 }
 
 size_t
 mos6502_inc_abs8_rx(int8_t value)
 {
-	return mos6502_write(0xf6) + mos6502_write(value);
+	return mos6502_write((int8_t) 0xf6) + mos6502_write(value);
 }
 
 size_t
 mos6502_inc_abs16(int16_t value)
 {
-	return mos6502_write(0xee) + mos6502_write(value);
+	return mos6502_write((int8_t) 0xee) + mos6502_write(value);
 }
 
 size_t
 mos6502_inc_abs16_rx(int16_t value)
 {
-	return mos6502_write(0xfe) + mos6502_write(value);
+	return mos6502_write((int8_t) 0xfe) + mos6502_write(value);
 }
