+++ src/mos6502/mos6502_ldx.c
@@ -16,30 +16,30 @@
 size_t
 mos6502_ldx_imm8(int8_t value)
 {
-	return mos6502_write(0xa2) + mos6502_write(value);
+	return mos6502_write((int8_t) 0xa2) + mos6502_write(value);
 }
 
 size_t
 mos6502_ldx_abs8(int8_t value)
 {
-	return mos6502_write(0xa6) + mos6502_write(value);
+	return mos6502_write((int8_t) 0xa6) + mos6502_write(value);
 }
 
 size_t
 mos6502_ldx_abs8_ry(int8_t value)
 {
-	return mos6502_write(0xb6) + mos6502_write(value);
+	return mos6502_write((int8_t) 0xb6) + mos6502_write(value);
 }
 
 size_t
 mos6502_ldx_abs16(int16_t value)
 {
-	return mos6502_write(0xae) + mos6502_imm16(value);
+	return mos6502_write((int8_t) 0xae) + mos6502_imm16(value);
 }
 
 size_t
 mos6502_ldx_abs16_ry(int16_t value)
 {
-	return mos6502_write(0xbe) + mos6502_imm16(value);
+	return mos6502_write((int8_t) 0xbe) + mos6502_imm16(value);
 }
