+++ src/z80/z80_add.c
@@ -17,28 +17,28 @@ size_t
 z80_add_r8(Z80_REG reg)
 {
 	if (z80_idx(reg))
-		return z80_i(reg) + z80_b(0x84, reg);
+		return z80_i(reg) + z80_b((int8_t) 0x84, reg);
 	else
-		return z80_rrr(0x80, 0, reg);
+		return z80_rrr((int8_t) 0x80, 0, reg);
 }
 
 size_t
 z80_add_imm8(int8_t value)
 {
-	return z80_write(0xc6) + z80_write(value);
+	return z80_write((int8_t) 0xc6) + z80_write(value);
 }
 
 size_t
 z80_add_rhl()
 {
-	return z80_write(0x86);
+	return z80_write((int8_t) 0x86);
 }
 
 size_t
 z80_add_ir16_off8(Z80_REG reg,
                   int8_t off)
 {
-	return z80_i(reg) + z80_write(0x86) + z80_write(off);
+	return z80_i(reg) + z80_write((int8_t) 0x86) + z80_write(off);
 }
 
 size_t
